tools/requirements.txt
numpy
Pillow
tflite-runtime
open-clip-torch
torch
torchvision
pytest
